Extract SetupCTAContent component from AudienceSegmentationSetupCTAWidget.
