+++ assets/js/modules/analytics-4/components/audience-segmentation/dashboard/AudienceSegmentationSetupCTAWidget/index.js
@@ -1,5 +1,5 @@
 /**
- * KeyMetricsSetupCTAWidget component.
+ * AudienceSegmentationSetupCTAWidget component.
  *
  * Site Kit by Google, Copyright 2024 Google LLC
  *
@@ -32,9 +32,6 @@ import { Fragment, useCallback, useState } from '@wordpress/element';
  * Internal dependencies
  */
 import { useDispatch, useSelect } from 'googlesitekit-data';
-import BannerGraphicsSVGDesktop from '../../../../../../../svg/graphics/audience-segmentation-setup-desktop.svg';
-import BannerGraphicsSVGTablet from '../../../../../../../svg/graphics/audience-segmentation-setup-tablet.svg';
-import BannerGraphicsSVGMobile from '../../../../../../../svg/graphics/audience-segmentation-setup-mobile.svg';
 import whenActive from '../../../../../../util/when-active';
 import { CORE_FORMS } from '../../../../../../googlesitekit/datastore/forms/constants';
 import { CORE_USER } from '../../../../../../googlesitekit/datastore/user/constants';
@@ -45,13 +42,6 @@ import {
 	AUDIENCE_SEGMENTATION_SETUP_FORM,
 } from '../../../../datastore/constants';
 import { SETTINGS_VISITOR_GROUPS_SETUP_SUCCESS_NOTIFICATION } from '../../settings/SettingsCardVisitorGroups/SetupSuccess';
-import { Button, SpinnerButton } from 'googlesitekit-components';
-import { Cell, Grid, Row } from '../../../../../../material-components';
-import {
-	BREAKPOINT_SMALL,
-	BREAKPOINT_TABLET,
-	useBreakpoint,
-} from '../../../../../../hooks/useBreakpoint';
 import useViewContext from '../../../../../../hooks/useViewContext';
 import {
 	AdminMenuTooltip,
@@ -62,16 +52,13 @@ import { withWidgetComponentProps } from '../../../../../../googlesitekit/widget
 import { WEEK_IN_SECONDS } from '../../../../../../util';
 import useEnableAudienceGroup from '../../../../hooks/useEnableAudienceGroup';
 import AudienceErrorModal from '../AudienceErrorModal';
+import SetupCTAContent from './SetupCTAContent';
 
 export const AUDIENCE_SEGMENTATION_SETUP_CTA_NOTIFICATION =
 	'audience_segmentation_setup_cta-notification';
 
 function AudienceSegmentationSetupCTAWidget( { Widget, WidgetNull } ) {
 	const viewContext = useViewContext();
-
-	const breakpoint = useBreakpoint();
-	const isMobileBreakpoint = breakpoint === BREAKPOINT_SMALL;
-	const isTabletBreakpoint = breakpoint === BREAKPOINT_TABLET;
 
 	const { invalidateResolution } = useDispatch( CORE_NOTIFICATIONS );
 
@@ -213,109 +200,14 @@ function AudienceSegmentationSetupCTAWidget( { Widget, WidgetNull } ) {
 		return null;
 	}
 
-	// TODO: We need to refactor this and the ConsentModeSetupCTAWidget to avoid this duplicate inlining of the widget context and area structure,
-	// and to ensure only one of these setup CTAs is shown at a time. This will be handled in a subsequent issue.
 	return (
-		<div className="googlesitekit-widget-context">
-			<Grid className="googlesitekit-widget-area">
-				<Row>
-					<Cell size={ 12 }>
-						<Widget
-							noPadding
-							className="googlesitekit-audience-segmentation-setup-cta-widget"
-						>
-							<Grid collapsed>
-								<Row>
-									<Cell
-										smSize={ 6 }
-										mdSize={ 8 }
-										lgSize={ 7 }
-										className="googlesitekit-widget-audience-segmentation-primary-cell"
-									>
-										<div className="googlesitekit-widget-audience-segmentation-text__wrapper">
-											<h3 className="googlesitekit-publisher-win__title">
-												{ __(
-													'Learn how different types of visitors interact with your site',
-													'google-site-kit'
-												) }
-											</h3>
-											<p>
-												{ __(
-													'Understand what brings new visitors to your site and keeps them coming back. Site Kit can now group your site visitors into relevant segments like "new" and "returning". To set up these new groups, Site Kit needs to update your Google Analytics property.',
-													'google-site-kit'
-												) }
-											</p>
-										</div>
-										<div className="googlesitekit-widget-audience-segmentation-actions__wrapper">
-											<Fragment>
-												<SpinnerButton
-													className="googlesitekit-audience-segmentation-cta-button"
-													onClick={ onEnableGroups }
-													isSaving={ isSaving }
-												>
-													{ isSaving
-														? __(
-																'Enabling groups',
-																'google-site-kit'
-														  )
-														: __(
-																'Enable groups',
-																'google-site-kit'
-														  ) }
-												</SpinnerButton>
-												<Button
-													tertiary
-													onClick={
-														handleDismissClick
-													}
-												>
-													{ dismissCount < 1
-														? __(
-																'Maybe later',
-																'google-site-kit'
-														  )
-														: __(
-																'Don’t show again',
-																'google-site-kit'
-														  ) }
-												</Button>
-											</Fragment>
-										</div>
-									</Cell>
-									{ ! isMobileBreakpoint &&
-										! isTabletBreakpoint && (
-											<Cell
-												alignBottom
-												className="googlesitekit-widget-audience-segmentation-svg__wrapper"
-												smSize={ 6 }
-												mdSize={ 3 }
-												lgSize={ 5 }
-											>
-												<BannerGraphicsSVGDesktop />
-											</Cell>
-										) }
-									{ isTabletBreakpoint && (
-										<Cell
-											className="googlesitekit-widget-audience-segmentation-svg__wrapper"
-											mdSize={ 8 }
-										>
-											<BannerGraphicsSVGTablet />
-										</Cell>
-									) }
-									{ isMobileBreakpoint && (
-										<Cell
-											className="googlesitekit-widget-audience-segmentation-svg__wrapper"
-											smSize={ 8 }
-										>
-											<BannerGraphicsSVGMobile />
-										</Cell>
-									) }
-								</Row>
-							</Grid>
-						</Widget>
-					</Cell>
-				</Row>
-			</Grid>
+		<Fragment>
+			<SetupCTAContent
+				Widget={ Widget }
+				onEnableGroups={ onEnableGroups }
+				isSaving={ isSaving }
+				handleDismissClick={ handleDismissClick }
+			/>
 			{ ( showErrorModal || hasOAuthError ) && (
 				<AudienceErrorModal
 					hasOAuthError={ hasOAuthError }
@@ -329,7 +221,7 @@ function AudienceSegmentationSetupCTAWidget( { Widget, WidgetNull } ) {
 					}
 				/>
 			) }
-		</div>
+		</Fragment>
 	);
 }
